requirements.txt
requests 
python-dotenv
websockets
kafka-python
psycopg2-binary